fix(test): self-consistent fixture in table replace test; stop pulling prisma into pure-logic suites
Two issues from Emily's local run:

- carbon_report_table replace test: the "updated ledger" fixture only
  overrode scopeSubtotals/totalCo2eKg while entries kept the old co2e —
  the detail row legitimately still printed 1,235,000. The replace logic
  was correct; the fixture wasn't self-consistent. Entries now carry the
  new value too
- Worker force-exit warning: carbon_articulation.service imported
  parseActivityQuantity from carbon_calculation.service, which pulls
  EmissionFactorRepo → real prisma (pg Pool) into every suite that
  touches articulation. The parser is a zero-dependency pure function —
  extracted to src/lib/carbon_quantity.ts; the calculation service
  re-exports it so existing importers are unchanged, and pure-logic
  suites no longer open a DB pool
