+++ src/services/carbon_calculation.service.ts
@@ -16,6 +16,7 @@ import {
   CarbonPendingReasonEnum,
 } from "@/constants/carbon_calculation";
 import { activityDedupeKey } from "@/lib/carbon_inventory";
+import { parseActivityQuantity } from "@/lib/carbon_quantity";
 import {
   IActivityRecord,
   IComputedLedger,
@@ -52,18 +53,9 @@ const defaultFactorLookup: IFactorLookup = {
   },
 };
 
-/**
- * Info: (20260716 - Emily) 數量字串決定性解析:全形轉半形、去千分位/空白;
- * 合法格式僅「非負十進位數」;其餘(科學記號/負數/文字)回 null → 待補,絕不猜。
- */
-export const parseActivityQuantity = (raw: string): string | null => {
-  const halfWidth = raw.replace(/[０-９．]/g, (ch) =>
-    ch === "．" ? "." : String.fromCharCode(ch.charCodeAt(0) - 0xfee0),
-  );
-  const cleaned = halfWidth.replace(/[,\s，]/g, "");
-  if (!/^\d+(\.\d+)?$/.test(cleaned)) return null;
-  return cleaned;
-};
+// Info: (20260720 - Emily) 數量解析抽至零依賴 lib(#6520 純邏輯呼叫端不得連動 prisma);
+// Info: (20260720 - Emily) re-export 維持既有呼叫端不變
+export { parseActivityQuantity };
 
 // Info: (20260716 - Emily) 係數單位正規化:完整字串或括號內代碼比對別名表;無法正規化回 null
 export const normalizeCoefficientUnit = (
